split inspector.rs
