+++ crates/inspector/src/lib.rs
@@ -9,6 +9,9 @@ extern crate alloc as std;
 mod eip3155;
 mod gas;
 mod inspector;
+pub mod inspector_context;
+pub mod inspector_frame;
+mod journal;
 mod noop;
 
 pub use inspector::*;
